Step 4

+++ bookStore/src/org/vermeg/bookstore/model/User.java
@@ -35,7 +35,7 @@ public class User {
 	public long getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public double getCin() {
